Use local terminal-notifier when present
The terminal-notifier executable included in the vendor directory is not
easily upgradeable, and certain users may wish to have more control over
which version of terminal-notifier is used (for instance to take
advantage of new -appIcon and -contentImage features). If a user needs
more control over terminal-notifier, this commit allows them to use
whatever version they choose to install locally via RubyGems or
Homebrew.