Snoar Toast appID option not being set when passed
A bug I encountered was that if I would pass `appID` option for windows toast, it would be removed and not set as an option for Snoar Toast.  
For it to work I had to use `appName` option then it worked perfectly fine.  
I could've created PR for README to change the API description. 
However, code change will be more consistent for the API that `node-notifier` is using.